fix swap offset

+++ test/mantle/MarginCollateralSwap.t.sol
@@ -6,7 +6,7 @@ import "./DeltaSetup.f.sol";
 contract MarginCollateralSwapTest is DeltaSetup {
     uint256 DEFAULT_IR_MODE = 2; // variable
 
-    function test_margin_mantle_collateral_exact_in(uint8 lenderId) external /** address user, uint8 lenderId */ {
+    function test_margin_mantle_collateral_exact_in(uint16 lenderId) external /** address user, uint16 lenderId */ {
         address user = testUser;
         vm.assume(user != address(0) && lenderId < 2);
         address asset = USDC;
@@ -50,7 +50,7 @@ contract MarginCollateralSwapTest is DeltaSetup {
         assertApproxEqAbs(14984996, balance, 1);
     }
 
-    function test_margin_mantle_collateral_exact_in_multi(uint8 lenderId) external {
+    function test_margin_mantle_collateral_exact_in_multi(uint16 lenderId) external {
         address user = testUser;
         vm.assume(user != address(0) && lenderId < 2);
         address asset = USDC;
@@ -94,7 +94,7 @@ contract MarginCollateralSwapTest is DeltaSetup {
         assertApproxEqAbs(14945822, balance, 1);
     }
 
-    function test_margin_mantle_collateral_exact_out(uint8 lenderId) external {
+    function test_margin_mantle_collateral_exact_out(uint16 lenderId) external {
         address user = testUser;
         vm.assume(user != address(0) && lenderId < 2);
         address asset = USDC;
@@ -138,7 +138,7 @@ contract MarginCollateralSwapTest is DeltaSetup {
         assertApproxEqAbs(amountOut, balance, 1);
     }
 
-    function test_margin_mantle_collateral_exact_out_multi(uint8 lenderId) external {
+    function test_margin_mantle_collateral_exact_out_multi(uint16 lenderId) external {
         address user = testUser;
         vm.assume(user != address(0) && lenderId < 2);
         address asset = USDC;
@@ -182,7 +182,7 @@ contract MarginCollateralSwapTest is DeltaSetup {
         assertApproxEqAbs(amountOut, balance, 1);
     }
 
-    function test_margin_mantle_collateral_all_in(uint8 lenderId) external {
+    function test_margin_mantle_collateral_all_in(uint16 lenderId) external {
         address user = testUser;
         vm.assume(user != address(0) && lenderId < 2);
         address asset = USDC;
@@ -231,7 +231,7 @@ contract MarginCollateralSwapTest is DeltaSetup {
 
     /** TEST FOR V2 CALLBACKS */
 
-    function test_margin_mantle_collateral_exact_in_v2(uint8 lenderId) external /** address user, uint8 lenderId */ {
+    function test_margin_mantle_collateral_exact_in_v2(uint16 lenderId) external /** address user, uint16 lenderId */ {
         address user = testUser;
         vm.assume(user != address(0) && lenderId < 2);
         address asset = USDC;
@@ -275,7 +275,7 @@ contract MarginCollateralSwapTest is DeltaSetup {
         assertApproxEqAbs(14909728, balance, 1);
     }
 
-    function test_margin_mantle_collateral_exact_in_multi_v2(uint8 lenderId) external {
+    function test_margin_mantle_collateral_exact_in_multi_v2(uint16 lenderId) external {
         address user = testUser;
         vm.assume(user != address(0) && lenderId < 2);
         address asset = USDC;
@@ -319,7 +319,7 @@ contract MarginCollateralSwapTest is DeltaSetup {
         assertApproxEqAbs(14884120, balance, 1);
     }
 
-    function test_margin_mantle_collateral_exact_out_v2(uint8 lenderId) external {
+    function test_margin_mantle_collateral_exact_out_v2(uint16 lenderId) external {
         address user = testUser;
         vm.assume(user != address(0) && lenderId < 2);
         address asset = USDC;
@@ -363,7 +363,7 @@ contract MarginCollateralSwapTest is DeltaSetup {
         assertApproxEqAbs(amountOut, balance, 1);
     }
 
-    function test_margin_mantle_collateral_exact_out_multi_v2(uint8 lenderId) external {
+    function test_margin_mantle_collateral_exact_out_multi_v2(uint16 lenderId) external {
         address user = testUser;
         vm.assume(user != address(0) && lenderId < 2);
         address asset = USDC;
@@ -407,7 +407,7 @@ contract MarginCollateralSwapTest is DeltaSetup {
         assertApproxEqAbs(amountOut, balance, 1);
     }
 
-    function test_margin_mantle_collateral_all_in_v2(uint8 lenderId) external {
+    function test_margin_mantle_collateral_all_in_v2(uint16 lenderId) external {
         address user = testUser;
         vm.assume(user != address(0) && lenderId < 2);
         address asset = USDC;
